requirements.txt
dash==2.24.1
dash-bootstrap-components==1.4.2
numpy==1.27.4
pandas==2.1.1
scikit-learn==1.7.1
mlflow==2.7.0
python-dotenv==1.0.0
pytest==8.4.2
joblib==1.3.2
gunicorn==21.2.0
